module/requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dm-tree==0.1.8
exceptiongroup==1.2.0
flatbuffers==23.5.26
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
openai==1.10.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pillow==10.2.0
playwright==1.35.0
protobuf==4.23.4
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
pyee==9.0.4
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
selectolax==0.3.14
selenium==4.15.2
six==1.16.0
slack-sdk==3.21.3
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.30.0
streamlit-option-menu==0.3.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
